+++ pyaud_plugins/_plugins/__init__.py
@@ -2,6 +2,6 @@
 pyaud_plugins._plugins
 ======================
 """
-from . import action, audit, fix, fix_file, parametrize, write
+from . import action, audit, fix, parametrize, write
 
-__all__ = ["action", "audit", "fix", "fix_file", "parametrize", "write"]
+__all__ = ["action", "audit", "fix", "parametrize", "write"]